Player profile (#221)
* Creating Player Profile prefab + new Leaderboards prefab
* Adding Filters and new shell
* Adding Controller Support for Filter views
* Updating the Leaderboard plugin to remove old controller variables, no longer needed components, refactoring a bit
* Updating Leaderboard Scene
* Adding theme handling in playerauthentication and leaderboards
* Fixing ThemeHelper bug + alignment issues
* Adding exception handling for get
* Removing game save from buildpackage.ps1 until fixed
* Fix Leaderboard Enable/Disable bug
* Fixing Sign In bug when the user cancels sign in
* Handling failed leaderboard requests gracefully.
* Fixing leaderboard bugs and playerauth ID position 
* Deleting old prefabs 
* Add LeaderboardHelper to allow developers to use the enum to socialgroup conversion instead of looking for the transition themselves.
* Moving LeaderbaordHelper to LeaderboardTypes

+++ Build/BuildPackage.ps1
@@ -17,62 +17,66 @@ if(!$unity)
 }
 
 $projectPath = Resolve-Path (Join-Path $PSScriptRoot ..)
-
-$gameSaveAssetsPath = "Assets\Xbox Live\GameSave"
-$tempgameSavePackagePath = Join-Path 'Assets\Xbox Live\' GameSave.unitypackage
-$gameSavePackagePath = Join-Path 'Assets\Xbox Live\GameSave\' GameSave.unitypackage
-Remove-Item $gameSavePackagePath -ErrorAction SilentlyContinue
-
-Write-Host "Moving Game Save Readme.txt into a temporary folder"
 $externalFolder = Resolve-Path (Join-Path $projectPath ..)
-$tempGameSaveFolder = New-Item (Join-Path $externalFolder 'tempGameSave') -type directory -force
-Move-Item (Resolve-Path (Join-Path $gameSaveAssetsPath 'README.txt')) -Destination $tempGameSaveFolder
 
-Write-Host "Exporting Xbox Live Game Save Unity Plugin to " -NoNewline
-Write-Host $gameSavePackagePath -ForegroundColor Green
-Write-Host "$($unity) -batchmode -logFile '$($logFile)' -projectPath '$($projectPath)' -exportPackage '$($gameSavePackagePath)' '$($gameSavePackagePath)' -quit"
-. $unity -batchmode -logFile "$logFile" -projectPath "$projectPath" -exportPackage "$gameSaveAssetsPath" "$tempgameSavePackagePath" -quit
-$global:unityProcess = Get-Process Unity | Sort-Object StartTime | Select-Object -Last 1 
+if (0)
+{
+    $gameSaveAssetsPath = "Assets\Xbox Live\GameSave"
+    $tempgameSavePackagePath = Join-Path 'Assets\Xbox Live\' GameSave.unitypackage
+    $gameSavePackagePath = Join-Path 'Assets\Xbox Live\GameSave\' GameSave.unitypackage
+    Remove-Item $gameSavePackagePath -ErrorAction SilentlyContinue
 
-# Wait for it to complete exporting the package
-$unityProcess | Wait-Process -Timeout 120 -ErrorAction SilentlyContinue
+    Write-Host "Moving Game Save Readme.txt into a temporary folder"
+    $tempGameSaveFolder = New-Item (Join-Path $externalFolder 'tempGameSave') -type directory -force
+    Move-Item (Resolve-Path (Join-Path $gameSaveAssetsPath 'README.txt')) -Destination $tempGameSaveFolder
 
-if(!$unityProcess.HasExited)
-{
-    Write-Warning "Unity (PID $($unityProcess.Id)) seems to be taking a long time to generate the package.  Check the log file for details to see what's happening."
-    Write-Warning "Log: $logFile"
+    Write-Host "Exporting Xbox Live Game Save Unity Plugin to " -NoNewline
+    Write-Host $gameSavePackagePath -ForegroundColor Green
+    Write-Host "$($unity) -batchmode -logFile '$($logFile)' -projectPath '$($projectPath)' -exportPackage '$($gameSavePackagePath)' '$($gameSavePackagePath)' -quit"
+    . $unity -batchmode -logFile "$logFile" -projectPath "$projectPath" -exportPackage "$gameSaveAssetsPath" "$tempgameSavePackagePath" -quit
+    $global:unityProcess = Get-Process Unity | Sort-Object StartTime | Select-Object -Last 1 
+
+    # Wait for it to complete exporting the package
+    $unityProcess | Wait-Process -Timeout 120 -ErrorAction SilentlyContinue
+
+    if(!$unityProcess.HasExited)
+    {
+        Write-Warning "Unity (PID $($unityProcess.Id)) seems to be taking a long time to generate the package.  Check the log file for details to see what's happening."
+        Write-Warning "Log: $logFile"
+    }
+    elseif($unityProcess.ExitCode -ne 0)
+    {
+        Write-Warning "Log: $logFile"
+        Write-Error "Unity.exe failed to build game save package.  Exit Code: $($unityProcess.ExitCode)."
+    }
+    else
+    {
+        Write-Host "Game Save Package created successfully. "
+    }
+    Write-Host ""
 }
-elseif($unityProcess.ExitCode -ne 0)
-{
-    Write-Warning "Log: $logFile"
-    Write-Error "Unity.exe failed to build game save package.  Exit Code: $($unityProcess.ExitCode)."
-}
-else
-{
-    Write-Host "Game Save Package created successfully. "
-}
-Write-Host ""
 
 $packagePath = Join-Path $projectPath XboxLive.unitypackage
 Remove-Item $packagePath -ErrorAction SilentlyContinue
 
-Write-Host "Moving Game Save Scripts and Prefabs to a temporary folder."
-Copy-Item $gameSaveAssetsPath -Destination $tempGameSaveFolder -ErrorAction SilentlyContinue -recurse
-Remove-Item $gameSaveAssetsPath -recurse
+if (0)
+{
+    Write-Host "Moving Game Save Scripts and Prefabs to a temporary folder."
+    Copy-Item $gameSaveAssetsPath -Destination $tempGameSaveFolder -ErrorAction SilentlyContinue -recurse
+    Remove-Item $gameSaveAssetsPath -recurse
 
 
-Write-Host ""
+    Write-Host ""
 
-Write-Host "Moving Game Save plugin into the 'GameSave' folder within Assets."
-New-Item $gameSaveAssetsPath -type directory
-Copy-Item $tempgameSavePackagePath -Destination $gameSaveAssetsPath
-Remove-Item $tempgameSavePackagePath
+    Write-Host "Moving Game Save plugin into the 'GameSave' folder within Assets."
+    New-Item $gameSaveAssetsPath -type directory
+    Copy-Item $tempgameSavePackagePath -Destination $gameSaveAssetsPath
+    Remove-Item $tempgameSavePackagePath
 
-Write-Host "Moving Readme.txt back into the Game Save folder ..."
-Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'README.txt')) -Destination $gameSaveAssetsPath 
-
-
-Write-Host ""
+    Write-Host "Moving Readme.txt back into the Game Save folder ..."
+    Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'README.txt')) -Destination $gameSaveAssetsPath 
+    Write-Host ""
+}
 
 Write-Host "Moving Scenes into a temporary folder ..."
 $scenesAssetsPath = "Assets\Scenes"
@@ -112,11 +116,14 @@ else
 }
 Write-Host ""
 
-Write-Host "Moving Game Save scripts and prefabs back into the 'GameSave' folder within Assets."
-Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'GameSave\Scripts')) -Destination $gameSaveAssetsPath
-Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'GameSave\Prefabs')) -Destination $gameSaveAssetsPath 
-Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'GameSave\Examples')) -Destination $gameSaveAssetsPath 
-Remove-Item $tempGameSaveFolder -recurse
+if (0)
+{
+    Write-Host "Moving Game Save scripts and prefabs back into the 'GameSave' folder within Assets."
+    Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'GameSave\Scripts')) -Destination $gameSaveAssetsPath
+    Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'GameSave\Prefabs')) -Destination $gameSaveAssetsPath 
+    Move-Item (Resolve-Path (Join-Path $tempGameSaveFolder 'GameSave\Examples')) -Destination $gameSaveAssetsPath 
+    Remove-Item $tempGameSaveFolder -recurse
+}
 
 Write-Host "Moving Scenes back into the Scenes folder ..."
 Move-Item (Resolve-Path (Join-Path $tempScenesFolder 'Scenes.meta')) -Destination "Assets"
